requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
fastapi==0.109.2
python-dotenv==1.0.1
uvicorn==0.29.0
